requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3
python-multipart==0.0.6
psutil==5.9.6

# Auth & Security
pyjwt==2.8.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-gnupg==0.5.2
cryptography==41.0.7

# Database & Cache
redis==5.0.1
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

# AutoGen
pyautogen==0.2.0

# Docker management
docker==6.1.3

# GPU & Compute
nvidia-ml-py==12.535.133
py3nvml==0.2.7
numba==0.58.1

# Market Data & Trading
yfinance==0.2.31
alpha-vantage==2.3.1
finnhub-python==2.4.19
ccxt==4.1.32
twelvedata==1.2.15
pandas-ta==0.3.14b0
ta==0.10.2
aiohttp==3.9.0
matplotlib==3.8.1
seaborn==0.13.0
mplfinance==0.12.10b0

# Async utilities
aiofiles==23.2.1
asyncio==3.4.3

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Machine Learning (for advanced strategies)
numpy==1.24.4
pandas==2.1.3
scipy==1.11.4
scikit-learn==1.3.2

# Blockchain & DEX (temporarily disabled due to dependency conflicts)
# web3==6.11.3
# eth-account==0.9.0
aiohttp==3.9.0

# Additional Data Processing
python-dateutil==2.8.2

# Visualization & Monitoring
rich==13.7.1

# Development dependencies (move to requirements-dev.txt in production)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-timeout==2.3.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0